+++ server/src/app.module.ts
@@ -4,7 +4,7 @@ import { typeORMConfig } from './typeorm.config';
 import { MenuModule } from './menu/menu.module';
 import { ChoiceModule } from './choice/choice.module';
 import { OrderModule } from './order/order.module';
-import { PaymentMethodModule } from './payment-method/payment-method.module';
+import { PaymentMethodModule } from './paymentMethod/paymentMethod.module';
 
 @Module({
   imports: [
