--- server/src/payment-method/dto/create-payment-method.dto.ts
@@ -1,1 +1,0 @@
-export class CreatePaymentMethodDto {}
