--- server/src/payment-method/dto/update-payment-method.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreatePaymentMethodDto } from './create-payment-method.dto';
-
-export class UpdatePaymentMethodDto extends PartialType(CreatePaymentMethodDto) {}
