--- server/src/payment-method/payment-method.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { PaymentMethodController } from './payment-method.controller';
-import { PaymentMethodService } from './payment-method.service';
-
-describe('PaymentMethodController', () => {
-  let controller: PaymentMethodController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [PaymentMethodController],
-      providers: [PaymentMethodService],
-    }).compile();
-
-    controller = module.get<PaymentMethodController>(PaymentMethodController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
