--- server/src/payment-method/payment-method.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { PaymentMethodService } from './payment-method.service';
-import { PaymentMethodController } from './payment-method.controller';
-
-@Module({
-  controllers: [PaymentMethodController],
-  providers: [PaymentMethodService]
-})
-export class PaymentMethodModule {}
