--- server/src/payment-method/payment-method.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { PaymentMethodService } from './payment-method.service';
-
-describe('PaymentMethodService', () => {
-  let service: PaymentMethodService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [PaymentMethodService],
-    }).compile();
-
-    service = module.get<PaymentMethodService>(PaymentMethodService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
