+++ server/src/paymentMethod/paymentMethod.module.ts
@@ -1,0 +1,12 @@
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { Module } from '@nestjs/common';
+import { PaymentMethodService } from './paymentMethod.service';
+import { PaymentMethodController } from './paymentMethod.controller';
+import { PaymentMethod } from './entities/paymentMethod.entity';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([PaymentMethod])],
+  controllers: [PaymentMethodController],
+  providers: [PaymentMethodService],
+})
+export class PaymentMethodModule {}
