+++ server/src/paymentMethod/paymentMethod.service.ts
@@ -1,0 +1,16 @@
+import { Injectable } from '@nestjs/common';
+import { InjectRepository } from '@nestjs/typeorm';
+import { Repository } from 'typeorm';
+import { PaymentMethod } from './entities/paymentMethod.entity';
+
+@Injectable()
+export class PaymentMethodService {
+  constructor(
+    @InjectRepository(PaymentMethod)
+    private paymentMethodRepository: Repository<PaymentMethod>,
+  ) {}
+
+  async findAll() {
+    return await this.paymentMethodRepository.find();
+  }
+}
